task_1

+++ lesson_3/task_1/main.cpp
@@ -7,13 +7,13 @@ int main()
     int arr[10], num = 0;
     cout << "Enter the number you want to add to the elements of the array: ";
     cin >> num;
-    for (short int i = 0; i < 10; i++){
+    for (int i = 0; i < 10; i++){
         cout << "Enter element array number " << i + 1 << ":";
         cin >> arr[i];
         arr[i] += num;
     }
     cout << endl;
-    for (short int i = 0; i < 10; i++){
+    for (int i = 0; i < 10; i++){
         cout << "Element array number " << i + 1 << ":" << arr[i]<< endl;
     }
 }
